+++ src/Solver.cpp
@@ -20,7 +20,7 @@ int DIE_UP_RIGHT_Y;
 // Bin info
 int BIN_WIDTH;
 int BIN_HEIGHT;
-int BIN_MAX_UTIL;
+double BIN_MAX_UTIL;
 // Delay info
 double DISP_DELAY;
